requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==2.4.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
